scripts/policy-deep-research-agent/training/requirements.txt
accelerate>=0.34.2
bitsandbytes>=0.43.3
datasets>=2.19.0
numpy>=1.26.4
openenv-core>=0.1.1
pydantic>=2.0.0
python-dotenv>=1.0.1
torch>=2.3.0
transformers>=4.46.3
trl>=0.9.6
langchain>=0.2.0
langchain-core
langgraph
langchain-openai
agentlightning
filelock
